test: extract shared two-tokenization fixture
The atomic-overlay-swap and pin/drain proofs duplicated the whole
class+shard+import+bucket-capture setup (~45 lines each — the SonarCloud
duplication gate red). Both now use one fixture; content unchanged.
