+++ adapters/repos/db/inverted_reindex_atomic_overlay_swap_test.go
@@ -19,17 +19,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/go-openapi/strfmt"
-	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 
 	"github.com/weaviate/weaviate/adapters/repos/db/helpers"
 	"github.com/weaviate/weaviate/adapters/repos/db/lsmkv"
 	"github.com/weaviate/weaviate/entities/models"
-	"github.com/weaviate/weaviate/entities/schema"
-	"github.com/weaviate/weaviate/entities/storobj"
-	"github.com/weaviate/weaviate/entities/tokenizer"
-	enthnsw "github.com/weaviate/weaviate/entities/vectorindex/hnsw"
 )
 
 // TestAtomicOverlaySwap_BM25NeverSeesZeroCount is the deterministic,
@@ -119,62 +113,13 @@ func TestAtomicOverlaySwap_OldCodeIsRacy(t *testing.T) {
 // the first such observation, and the number of reads performed.
 func runAtomicOverlaySwapProof(t *testing.T) (sawBadOut bool, detailOut string, readsOut int64) {
 	ctx := testCtx()
-	const (
-		className   = "RetokWordRace"
-		fieldProp   = "alpha" // FIELD-tokenized: pre-swap content
-		wordProp    = "beta"  // WORD-tokenized: post-swap content
-		phrase      = "hello world"
-		filler      = "lorem ipsum"
-		numDocs     = 8
-		matchDocs   = 4 // docs carrying the queried phrase (the rest carry filler)
-		validCount  = matchDocs
-		hookSleepMs = 50
-	)
+	const hookSleepMs = 50
 
-	class := buildTwoTokenizationClass(className, fieldProp, wordProp)
-
-	shd, idx := testShardWithSettings(t, ctx, class, enthnsw.UserConfig{Skip: true},
-		false, false, false)
-	shard := shd.(*Shard)
-	t.Cleanup(func() { _ = shard.Shutdown(ctx) })
-
-	// Both props start at Inverted/BlockMax (UsingBlockMaxWAND:true).
-	for _, p := range []string{fieldProp, wordProp} {
-		require.Equal(t, lsmkv.StrategyInverted,
-			shard.store.Bucket(helpers.BucketSearchableFromPropNameLSM(p)).Strategy(),
-			"searchable bucket for %q must start at Inverted", p)
-	}
-
-	// Import N docs. The first matchDocs carry the queried phrase in BOTH
-	// props; the rest carry filler text. Keeping the queried terms out of
-	// 100% of docs avoids BM25's over-frequent-term IDF collapse, so the
-	// phrase reliably scores above zero and the matched docs are returned.
-	for i := 0; i < numDocs; i++ {
-		text := phrase
-		if i >= matchDocs {
-			text = filler
-		}
-		obj := &storobj.Object{
-			MarshallerVersion: 1,
-			Object: models.Object{
-				ID:    strfmt.UUID(uuid.NewString()),
-				Class: className,
-				Properties: map[string]interface{}{
-					fieldProp: text,
-					wordProp:  text,
-				},
-			},
-		}
-		require.NoError(t, shard.PutObject(ctx, obj))
-	}
-
-	// Capture the two distinct bucket identities up front so the proof can
-	// assert which bucket the resolver hands back at any instant.
-	fieldBucket := shard.store.Bucket(helpers.BucketSearchableFromPropNameLSM(fieldProp))
-	wordBucket := shard.store.Bucket(helpers.BucketSearchableFromPropNameLSM(wordProp))
-	require.NotNil(t, fieldBucket)
-	require.NotNil(t, wordBucket)
-	require.NotSame(t, fieldBucket, wordBucket, "field and word buckets must be distinct objects")
+	fx := setupTwoTokenizationShard(t, ctx, "RetokWordRace")
+	shard, idx := fx.shard, fx.idx
+	fieldBucket, wordBucket := fx.fieldBucket, fx.wordBucket
+	className, fieldProp, wordProp := fx.className, fx.fieldProp, fx.wordProp
+	phrase, validCount := fx.phrase, fx.matchDocs
 
 	// Baseline equivalence the proof hinges on — verified by emulating a
 	// query END-TO-END from a (tokenization, bucket) pair: tokenize the
@@ -311,70 +256,4 @@ func runAtomicOverlaySwapProof(t *testing.T) (sawBadOut bool, detailOut string, 
 		"post-swap: the (WORD tok, WORD bucket) pair must find the phrase docs")
 
 	return sawBad.Load(), badDetail, totalReads.Load()
-}
-
-// buildTwoTokenizationClass builds a class with one FIELD-tokenized and one
-// WORD-tokenized searchable text property, on Inverted/BlockMax searchable
-// buckets (the production default for a field→word retokenization).
-func buildTwoTokenizationClass(className, fieldProp, wordProp string) *models.Class {
-	vFalse := false
-	vTrue := true
-	mkProp := func(name, tok string) *models.Property {
-		return &models.Property{
-			Name:            name,
-			DataType:        schema.DataTypeText.PropString(),
-			Tokenization:    tok,
-			IndexFilterable: &vFalse,
-			IndexSearchable: &vTrue,
-		}
-	}
-	return &models.Class{
-		Class:             className,
-		VectorIndexConfig: enthnsw.NewDefaultUserConfig(),
-		InvertedIndexConfig: &models.InvertedIndexConfig{
-			CleanupIntervalSeconds: 60,
-			Bm25:                   &models.BM25Config{K1: 1.2, B: 0.75},
-			Stopwords:              &models.StopwordConfig{Preset: "none"},
-			IndexNullState:         true,
-			IndexPropertyLength:    true,
-			// Inverted (BlockMax) searchable buckets — the production default
-			// and the strategy a field→word searchable retokenization runs on.
-			UsingBlockMaxWAND: true,
-		},
-		Properties: []*models.Property{
-			mkProp(fieldProp, models.PropertyTokenizationField),
-			mkProp(wordProp, models.PropertyTokenizationWord),
-		},
-	}
-}
-
-// lookupCount emulates a keyword query END-TO-END from a (tokenization,
-// bucket) pair, the unit of consistency the FINALIZING-window fix protects:
-// it tokenizes the query the given way, looks every resulting term up in the
-// given searchable bucket via the real DocPointerWithScoreList primitive,
-// and returns the best per-term hit count. A CONSISTENT pair (query
-// tokenized the same way the bucket was indexed) finds the docs; a MIXED
-// pair (the bug) misses → 0.
-//
-// We drive the lookup directly off the (tok, bucket) pair rather than
-// through the full BM25 scoring pipeline so the proof isolates exactly the
-// pair-consistency property — the higher-level scorer adds IDF/limit
-// behavior irrelevant to the race and is exercised separately by the
-// integration BM25 suites.
-func lookupCount(ctx context.Context, tokenization string, bucket *lsmkv.Bucket, className, query string) int {
-	if bucket == nil {
-		return 0
-	}
-	terms := tokenizer.TokenizeForClass(tokenization, query, className)
-	best := 0
-	for _, term := range terms {
-		dp, err := bucket.DocPointerWithScoreList(ctx, []byte(term), 1)
-		if err != nil {
-			return -1
-		}
-		if len(dp) > best {
-			best = len(dp)
-		}
-	}
-	return best
 }
